Add ability to enqueue a task that does not require monitoring a return value
Add the ability to enqueue a task that does not require monitoring a
return value.  std::future<T> on destruction is expected to wait() for a
thread to complete before completing destruction (btw, MSVS2103
std::future/std::ansync does not behave according to C++ Standard).  If
a temporary future is created (not captured), the calling thread should
block waiting for the thread to exit and complete the promise.  An
alternate enqueueAndDetach method was added to allow for invocations
that don't require the capturing of the std::future.